+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 - Use bract.core 0.5.1
   - [TODO] Use the GA version
 - [TODO] Fix commands visibility in CLI help message
-  - [TODO] In root CLI inducers move `read-context` ahead of `parse-args`
-  - [TODO] Add `:bract.cli/pre-parse-inducers` to run before parsing CLI args
+  - In root CLI inducers move `read-context` ahead of `parse-args`
+  - [TODO] Add `:bract.cli/pre-inducers` to run before parsing CLI args
 - Make `bract.cli.main/-main` return exit code using `:bract.core/exit-code`